requirements.txt
aiohttp==2.2.5
async-timeout==1.2.1
certifi==2017.7.27.1
chardet==3.0.4
cleverwrap==0.2.3.6
discord.py==0.16.8
futures==3.1.1
goslate==1.5.1
idna==2.6
multidict==3.1.3
packaging==16.8
pyparsing==2.2.0
requests==2.18.3
six==1.10.0
urllib3==1.22
websockets==3.3
yarl==0.12.0
